V0 migration model. No surgery. See details.
PR #694 had a lot of conflicts related to PR #711, and discussions
suggested avoiding rename of existing tables. Rather than cherry pick
and rebase, this commit replaces the work in #694.
