+++ src/spyglass/linearization/__init__.py
@@ -1,0 +1,1 @@
+from spyglass.linearization.merge import LinearizedPositionOutput
