+++ src/spyglass/linearization/v1/__init__.py
@@ -1,0 +1,6 @@
+from spyglass.linearization.v1.main import (
+    LinearizationParameters,
+    LinearizationSelection,
+    LinearizedPositionV1,
+    TrackGraph,
+)
