+++ src/spyglass/linearization/v1/main.py
@@ -174,7 +174,7 @@ class LinearizedPositionV1(SpyglassMixin, dj.Computed):
 
         self.insert1(key)
 
-        from ..position_linearization_merge import LinearizedPositionOutput
+        from ..merge import LinearizedPositionOutput
 
         part_name = to_camel_case(self.table_name.split("__")[-1])
