--- src/spyglass/position_linearization/__init__.py
@@ -1,3 +1,0 @@
-from spyglass.position_linearization.position_linearization_merge import (
-    LinearizedPositionOutput,
-)
