--- src/spyglass/position_linearization/v1/__init__.py
@@ -1,7 +1,0 @@
-from spyglass.common.common_position import NodePicker
-from spyglass.position_linearization.v1.linearization import (
-    LinearizationParameters,
-    LinearizationSelection,
-    LinearizedPositionV1,
-    TrackGraph,
-)
