+++ src/spyglass/utils/__init__.py
@@ -1,0 +1,4 @@
+from spyglass.utils.dj_merge_tables import _Merge
+from spyglass.utils.dj_mixin import SpyglassMixin
+
+__all__ = ["_Merge", "SpyglassMixin"]
