+++ src/spyglass/utils/dj_helper_fn.py
@@ -1,11 +1,70 @@
 """Helper functions for manipulating information from DataJoint fetch calls."""
 import inspect
 import os
+from typing import Type
 
 import datajoint as dj
 import numpy as np
 
 from .nwb_helper_fn import get_nwb_file
+
+
+def deprecated_factory(classes: list, old_module: str = "") -> list:
+    """Creates a list of classes and prints a warning when instantiated
+
+    Parameters
+    ---------
+    classes : list
+        list of tuples containing old_class, new_class
+
+    Returns
+    ------
+    list
+        list of classes that will print a warning when instantiated
+    """
+
+    if not isinstance(classes, list):
+        classes = [classes]
+
+    ret = [
+        _subclass_factory(old_name=c[0], new_class=c[1], old_module=old_module)
+        for c in classes
+    ]
+
+    return ret[0] if len(ret) == 1 else ret
+
+
+def _subclass_factory(
+    old_name: str, new_class: Type, old_module: str = ""
+) -> Type:
+    """Creates a subclass with a deprecation warning on __init__
+
+    Old class is a subclass of new class, so it will inherit all of the new
+    class's methods. Old class retains its original name and module. Use
+    __name__ to get the module name of the caller.
+
+    Usage: OldClass = _subclass_factory('OldClass', __name__, NewClass)
+    """
+
+    new_module = new_class().__class__.__module__
+
+    # Define the __call__ method for the new class
+    def init_override(self, *args, **kwargs):
+        print(
+            "Deprecation Warning: this class has been moved out of "
+            + f"{old_module}\n"
+            + f"\t{old_name} -> {new_module}.{new_class.__name__}"
+            + "\nPlease use the new location."
+        )
+        return super(self.__class__, self).__init__(*args, **kwargs)
+
+    class_dict = {
+        "__module__": old_module or new_class.__class__.__module__,
+        "__init__": init_override,
+        "_is_deprecated": True,
+    }
+
+    return type(old_name, (new_class,), class_dict)
 
 
 def dj_replace(original_table, new_values, key_column, replace_column):
